Refactor school logic to vertical slice
