+++ src/app/api/profiles/search/route.ts
@@ -8,11 +8,11 @@ import {
   type RelaxDimension,
   type RelaxSuggestion,
   normalizeDashboardFilters,
-} from "@/lib/dashboardFilters";
+} from "@/features/school/data/dashboardFilters";
 import {
   getSchoolLabel,
   SECTOR_INTEREST_LABELS,
-} from "@/lib/utSchoolsAndMajors";
+} from "@/features/school/data/utSchoolsAndMajors";
 
 const SUPABASE_URL = process.env.NEXT_PUBLIC_SUPABASE_URL!;
 const SERVICE_KEY = process.env.SUPABASE_SERVICE_ROLE_KEY!;
