+++ src/features/school/auth/email-domain.ts
