+++ src/features/school/components/landing/DepartmentMosaic.tsx
