+++ src/features/school/components/landing/Hero.tsx
