+++ src/features/school/components/landing/HowItWorks.tsx
