+++ src/features/school/components/landing/PublicFooter.tsx
