+++ src/features/school/components/landing/ValuesPillars.tsx
