+++ src/features/school/data/utSchoolsAndMajors.ts
