+++ src/features/school/registry/registry.ts
