+++ src/features/school/registry/types.ts
